requirements_kai.txt
absl-py==0.15.0
astunparse==1.6.3
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
dill==0.3.9
dm-tree==0.1.8
docker-pycreds==0.4.0
filelock==3.14.0
flatbuffers==1.12
fonttools==4.53.0
fsspec==2024.6.0
future==1.0.0
gast==0.3.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.32.0
h5py==2.10.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.0
Jinja2==3.1.4
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.6.3
minerl-navigate==1.1.0
mpmath==1.3.0
networkx==3.1
numpy==1.19.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.1
pandas==1.1.5
pillow==10.3.0
platformdirs==4.3.6
promise==2.3
protobuf==3.20.3
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.18.1
scipy==1.10.1
sentry-sdk==2.17.0
setproctitle==1.3.3
six==1.15.0
smmap==5.0.1
sympy==1.12.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.1
tensorflow-datasets==4.2.0
tensorflow-estimator==2.4.0
tensorflow-metadata==1.14.0
tensorflow-probability==0.12.1
termcolor==1.1.0
tifffile==2023.7.10
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.6
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wandb==0.18.5
Werkzeug==3.0.6
wrapt==1.12.1
zipp==3.20.2
